requirements.txt
beautifulsoup4==4.9.3
chardet==4.0.0
pymssql==2.2.2
PyMySQL==1.0.2
PyYAML==5.4.1
selenium==3.141.0
soupsieve==2.2.1
urllib3==1.26.7
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
